refactor: Move `Posthog.feature_flag` to `Client.feature_flag`
That's where it belongs
