+++ lib/posthog/client.ex
@@ -1,3 +1,5 @@
+require Logger
+
 defmodule Posthog.Client do
   @moduledoc """
   Low-level HTTP client for interacting with PostHog's API.
@@ -72,6 +74,13 @@ defmodule Posthog.Client do
       Posthog.Client.feature_flags("user_123", groups: %{team: "engineering"})
   """
 
+  alias Posthog.FeatureFlag
+
+  @typedoc """
+  Result of a PostHog operation.
+  """
+  @type result() :: {:ok, response()} | {:error, response() | term()}
+
   @typedoc """
   HTTP headers in the format expected by :hackney.
   """
@@ -129,6 +138,23 @@ defmodule Posthog.Client do
   """
   @type feature_flag_opts :: opts() | [send_feature_flag_event: boolean()]
 
+  @typedoc """
+  Cache key for the `$feature_flag_called` event.
+  """
+  @type cache_key() :: {:feature_flag_called, binary(), binary()}
+
+  @typep feature_flag_called_event_properties_key() ::
+           :"$feature_flag"
+           | :"$feature_flag_response"
+           | :"$feature_flag_id"
+           | :"$feature_flag_version"
+           | :"$feature_flag_reason"
+           | :"$feature_flag_request_id"
+           | :distinct_id
+  @typep feature_flag_called_event_properties() :: %{
+           feature_flag_called_event_properties_key() => any() | nil
+         }
+
   # Adds default headers to the request.
   #
   # ## Parameters
@@ -166,8 +192,7 @@ defmodule Posthog.Client do
       # Event with custom headers
       Posthog.Client.capture("login", "user_123", %{}, headers: [{"x-forwarded-for", "127.0.0.1"}])
   """
-  @spec capture(event(), distinct_id(), properties(), opts()) ::
-          {:ok, response()} | {:error, response() | term()}
+  @spec capture(event(), distinct_id(), properties(), opts()) :: result()
   def capture(event, distinct_id, properties \\ %{}, opts \\ []) when is_list(opts) do
     if Posthog.Config.enabled_capture?() do
       posthog_event = Posthog.Event.new(event, distinct_id, properties, opts)
@@ -195,8 +220,7 @@ defmodule Posthog.Client do
 
       Posthog.Client.batch(events, %{timestamp: DateTime.utc_now()})
   """
-  @spec batch([{event(), distinct_id(), properties()}], opts(), headers()) ::
-          {:ok, response()} | {:error, response() | term()}
+  @spec batch([{event(), distinct_id(), properties()}], opts(), headers()) :: result()
   def batch(events, opts) when is_list(opts) do
     batch(events, opts, headers(opts[:headers]))
   end
@@ -235,8 +259,7 @@ defmodule Posthog.Client do
         group_properties: %{company: %{industry: "tech"}}
       )
   """
-  @spec feature_flags(binary(), opts()) ::
-          {:ok, Posthog.FeatureFlag.flag_response()} | {:error, response() | term()}
+  @spec feature_flags(binary(), opts()) :: result()
   def feature_flags(distinct_id, opts) do
     case _decide_request(distinct_id, opts) do
       {:ok, response} ->
@@ -249,6 +272,132 @@ defmodule Posthog.Client do
       err ->
         err
     end
+  end
+
+  @doc """
+  Retrieves information about a specific feature flag for a given distinct ID.
+
+  ## Parameters
+
+    * `flag` - The name of the feature flag
+    * `distinct_id` - The unique identifier for the user
+    * `opts` - Optional parameters for the feature flag request
+
+  ## Examples
+
+      # Boolean feature flag
+      {:ok, flag} = Posthog.feature_flag("new-dashboard", "user_123")
+      # Returns: %Posthog.FeatureFlag{name: "new-dashboard", payload: true, enabled: true}
+
+      # Multivariate feature flag
+      {:ok, flag} = Posthog.feature_flag("pricing-test", "user_123")
+      # Returns: %Posthog.FeatureFlag{
+      #   name: "pricing-test",
+      #   payload: %{"price" => 99, "period" => "monthly"},
+      #   enabled: "variant-a"
+      # }
+  """
+  @spec feature_flag(binary(), binary(), feature_flag_opts()) :: result()
+  def feature_flag(flag, distinct_id, opts \\ []) do
+    with {:ok, response} <- _decide_request(distinct_id, opts),
+         enabled when not is_nil(enabled) <- response.feature_flags[flag] do
+      # Only capture if send_feature_flag_event is true (default)
+      if Keyword.get(opts, :send_feature_flag_event, true),
+        do:
+          capture_feature_flag_called_event(
+            distinct_id,
+            %{
+              "$feature_flag" => flag,
+              "$feature_flag_response" => enabled
+            },
+            response
+          )
+
+      {:ok, FeatureFlag.new(flag, enabled, Map.get(response.feature_flag_payloads, flag))}
+    else
+      {:error, _} = err -> err
+      nil -> {:error, :not_found}
+    end
+  end
+
+  @spec capture_feature_flag_called_event(
+          distinct_id(),
+          feature_flag_called_event_properties(),
+          map()
+        ) ::
+          :ok
+  defp capture_feature_flag_called_event(distinct_id, properties, response) do
+    # Create a unique key for this distinct_id and flag combination
+    cache_key = {:feature_flag_called, distinct_id, properties["$feature_flag"]}
+
+    # Check if we've seen this combination before using Cachex
+    case Cachex.exists?(Posthog.Application.cache_name(), cache_key) do
+      {:ok, false} ->
+        do_capture_feature_flag_called_event(cache_key, distinct_id, properties, response)
+
+      # Should be `{:error, :no_cache}` but Dyalixir is wrongly assuming that doesn't exist
+      {:error, _} ->
+        # Cache doesn't exist, let's capture the event PLUS notify user they should be initing it
+        do_capture_feature_flag_called_event(cache_key, distinct_id, properties, response)
+
+        Logger.error("""
+        [posthog] Cachex process `#{inspect(Posthog.Application.cache_name())}` is not running.
+
+        ➤ This likely means you forgot to include `posthog` as an application dependency (mix.exs):
+
+            Example:
+
+            extra_applications: [..., :posthog]
+
+
+        ➤ Or, add `Posthog.Application` to your supervision tree (lib/my_lib/application.ex).
+
+            Example:
+                {Posthog.Application, []}
+        """)
+
+      {:ok, true} ->
+        # Entry already exists, no need to do anything
+        :ok
+    end
+  end
+
+  @spec do_capture_feature_flag_called_event(
+          cache_key(),
+          distinct_id(),
+          feature_flag_called_event_properties(),
+          map()
+        ) :: :ok
+  defp do_capture_feature_flag_called_event(cache_key, distinct_id, properties, response) do
+    flag = properties["$feature_flag"]
+
+    properties =
+      if Map.has_key?(response, :flags) do
+        Map.merge(properties, %{
+          "$feature_flag_id" => response.flags[flag]["metadata"]["id"],
+          "$feature_flag_version" => response.flags[flag]["metadata"]["version"],
+          "$feature_flag_reason" => response.flags[flag]["reason"]["description"]
+        })
+      else
+        properties
+      end
+
+    properties =
+      if Map.get(response, :request_id) do
+        Map.put(properties, "$feature_flag_request_id", response.request_id)
+      else
+        properties
+      end
+
+    # Send the event to our server
+    # NOTE: Calling this with `Posthog.Client.capture/4` rather than `capture/4`
+    #       because mocks won't work properly unless we use the fully defined function
+    Posthog.Client.capture("$feature_flag_called", distinct_id, properties, [])
+
+    # Add new entry to cache using Cachex
+    Cachex.put(Posthog.Application.cache_name(), cache_key, true)
+
+    :ok
   end
 
   @doc false
